% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRCL_functions.R
\name{SRCL_1_initiate_neural_network}
\alias{SRCL_1_initiate_neural_network}
\title{Initiates a tailored monotonistc neural network}
\usage{
SRCL_1_initiate_neural_network(inputs, hidden, confounder = FALSE)
}
\arguments{
\item{inputs}{The number of exposures.}

\item{hidden}{Number of hidden nodes.}

\item{confounder}{Allows to control away a confounder (connected to the output layer)}
}
\description{
This function initiates a tailored monotonistc neural network.
}
\examples{
#See the example under SRCL_0_motivating_example

}
