% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRCL_functions.R
\name{SRCL_4_predict_risks}
\alias{SRCL_4_predict_risks}
\title{Layer-wise relevance propagation of the fitted monotonistic neural network}
\usage{
SRCL_4_predict_risks(X, model)
}
\arguments{
\item{X}{The exposure data}

\item{model}{The fitted the monotonistic neural network}
}
\description{
Calculates risk contributions using layer-wise relevance propagation of the fitted monotonistic neural network and data
}
\examples{
#See the example under SRCL_0_motivating_example
}
